refactor(architecture): split transport helper surface
