+++ docs/adr/0013-composable-session-capabilities.md
@@ -218,7 +218,7 @@ narrative of how the cutover landed.
 - **C-Z. AST-based delegate-surface regression guard** — introduced by PR [#885](https://github.com/teng-lin/notebooklm-py/pull/885) (`f48d4b9`). To prevent erosion of the session-facade decoupling contract, a regression test in `tests/unit/test_session_compat_delegates.py` uses AST analysis to enforce that eight legacy/compatibility methods on the `Session` facade remain simple delegate forwards (6 `RpcExecutor`-adjacent and 2 `AuthRefreshCoordinator`-adjacent methods, each restricted to a maximum of 3 statements).
 - **C-AA. Localized `DrainHookRegistration` Protocol** — The `DrainHookRegistration` protocol has been retired from the general `_session_contracts.py` surface and is now local to the `ArtifactsRuntime` in `_artifacts.py`.
 - **C-AB. Narrow executor host Protocols after bridge retirement** — The session-shrink arc narrowed `RpcOwner` after the legacy `Session` private-attribute shims were retired. `RpcOwner` no longer declares `_timeout`, `_refresh_callback`, `_refresh_retry_delay`, or `_http_client`; `RpcExecutor` receives those values through constructor-time providers or the concrete `Kernel`.
-- **C-AC. Retire the authed-transport Adapter** — The middleware chain terminal now calls `Kernel.post` directly through `Session._authed_post_chain_terminal`. `_authed_transport.py` retains transport request types, exceptions, `Retry-After` parsing, and streaming helpers, while `_transport_errors.py` owns raw `Kernel.post` error mapping. This removed the last `_AuthedTransportHost` Protocol and the shallow `AuthedTransport` Adapter seam.
+- **C-AC. Retire the legacy transport Adapter** — The middleware chain terminal now calls `Kernel.post` directly through `Session._authed_post_chain_terminal`. Request construction lives in `_request_types.py`, transport exceptions and `Retry-After` parsing live in `_transport_errors.py`, and size-capped streaming lives in `_streaming_post.py`. This removed the last legacy host Protocol and the shallow Adapter seam.
 
 ## Alternatives considered
