+++ tests/_lint/test_no_session_compat_bridges.py
@@ -107,8 +107,10 @@ CARVE_OUT_MODULES: frozenset[str] = frozenset(
         "src/notebooklm/_cookie_persistence.py",
         "src/notebooklm/_polling_registry.py",
         "src/notebooklm/_reqid_counter.py",
-        "src/notebooklm/_authed_transport.py",
         "src/notebooklm/_rpc_executor.py",
+        "src/notebooklm/_request_types.py",
+        "src/notebooklm/_streaming_post.py",
+        "src/notebooklm/_transport_errors.py",
         "src/notebooklm/_middleware_auth_refresh.py",
         "src/notebooklm/_middleware_chain.py",
         "src/notebooklm/_middleware_drain.py",
