+++ tests/integration/test_session_integration.py
@@ -519,7 +519,7 @@ class TestBuildUrlHL:
 
     @staticmethod
     def _snapshot_for(core):
-        from notebooklm._authed_transport import AuthSnapshot
+        from notebooklm._request_types import AuthSnapshot
 
         return AuthSnapshot(
             csrf_token=core.auth.csrf_token,
