+++ tests/unit/test_retry_middleware.py
@@ -39,10 +39,10 @@ import pytest
 # pytest puts ``tests/`` on ``sys.path``; ``_fixtures.chain`` is the canonical
 # import path documented in ``tests/_fixtures/__init__.py``.
 from _fixtures.chain import make_request
-from notebooklm._authed_transport import TransportRateLimited, TransportServerError
 from notebooklm._client_metrics import ClientMetrics
 from notebooklm._middleware import NextCall, RpcRequest, RpcResponse, build_chain
 from notebooklm._middleware_retry import RetryMiddleware
+from notebooklm._transport_errors import TransportRateLimited, TransportServerError
 
 
 def _recording_sleep() -> tuple[Callable[[float], Awaitable[None]], list[float]]:
